ArchGemini/arch-gemini/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "ArchGemini Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "httpx",
    "Pillow",
    "python-multipart",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "core",
    "services",
]
include = [
    "app.py",
]
